+++ crates/ruma-common/tests/events/event_enums.rs
@@ -6,7 +6,9 @@ use ruma_common::{
 use serde_json::{from_value as from_json_value, json};
 
 use ruma_common::events::{
-    call::{answer::CallAnswerEventContent, SessionDescription, SessionDescriptionType},
+    call::{
+        answer::CallAnswerEventContent, AnswerSessionDescription, AnswerSessionDescriptionType,
+    },
     AnyMessageLikeEvent, OriginalMessageLikeEvent,
 };
 
@@ -41,8 +43,8 @@ fn deserialize_message_event() {
             .unwrap(),
         AnyMessageLikeEvent::CallAnswer(MessageLikeEvent::Original(OriginalMessageLikeEvent {
             content: CallAnswerEventContent {
-                answer: SessionDescription {
-                    session_type: SessionDescriptionType::Answer,
+                answer: AnswerSessionDescription {
+                    session_type: AnswerSessionDescriptionType::Answer,
                     sdp,
                     ..
                 },